+++ src/constants/queryKeys.ts
@@ -1,1 +1,2 @@
 export const QUERY_KEY_STATION_DETAILS = 'stationDetails';
+export const QUERY_KEY_STATION_SUMMARY = 'stationSummary';
