--- src/hooks/stations/useFetchStations.ts
@@ -1,52 +1,0 @@
-import axios from 'axios';
-
-import { useEffect, useState } from 'react';
-
-import { useExternalState } from '@util/external-state';
-
-import { stationSummaryListStore } from '@store/stationSummaryListStore';
-
-import { ROWS_PER_PAGE } from '@constant';
-import { BASE_URL } from '@constant/url';
-
-import type { StationSummary } from '@type';
-
-interface StationsResponse {
-  lastPage: number;
-  elements: StationSummary[];
-}
-
-export const useFetchStations = (token: string, page: number) => {
-  const [lastPage, setLastPage] = useState(1);
-  const [stationSummaryList, setStationSummaryList] =
-    useExternalState<StationSummary[]>(stationSummaryListStore);
-
-  const headers = {
-    headers: {
-      Authorization: `Bearer ${token}`,
-      'Content-Type': 'application/json',
-    },
-  };
-
-  useEffect(() => {
-    const fetchStations = async () => {
-      try {
-        const {
-          data: { lastPage, elements: stationSummaryList },
-        } = await axios.get<StationsResponse>(
-          `${BASE_URL}/stations?page=${page}&size=${ROWS_PER_PAGE}`,
-          headers
-        );
-
-        setLastPage(lastPage);
-        setStationSummaryList(stationSummaryList);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
-    fetchStations();
-  }, []);
-
-  return { lastPage, stationSummaryList };
-};
