+++ src/stores/pageStore.ts
@@ -1,0 +1,3 @@
+import { store } from "@util/external-state";
+
+export const currentPage = store<number>(1);
